+++ src/main/java/ch/gabrielegli/flugverwaltung/service/AirportService.java
@@ -27,7 +27,9 @@ public class AirportService {
     @GET
     @Path("list")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response getAllAirports(@QueryParam("sort") String sort) {
+    public Response getAllAirports(@CookieParam("userRole") String userRole, @QueryParam("sort") String sort) {
+        if (!isUserPermitted(userRole, new String[]{"admin", "user"})) return Response.status(401).build();
+
         List<Airport> airports = DataHandler.readAllAirport();
         int status = 200;
 
@@ -67,10 +69,13 @@ public class AirportService {
     @Path("get")
     @Produces(MediaType.APPLICATION_JSON)
     public Response getOneAirport(
+            @CookieParam("userRole") String userRole,
             @QueryParam("uuid")
             @NotEmpty
             @Pattern(regexp = "[0-9a-fA-F]{8}-([0-9a-fA-F]{4}-){3}[0-9a-fA-F]{12}") String uuid
     ) {
+        if (!isUserPermitted(userRole, new String[]{"admin", "user"})) return Response.status(401).build();
+
         Airport airport = DataHandler.readAirportByUUID(uuid);
         int status = 200;
 
@@ -94,11 +99,14 @@ public class AirportService {
     @Path("delete")
     @Produces(MediaType.TEXT_PLAIN)
     public Response deleteOneAirport(
+            @CookieParam("userRole") String userRole,
             @QueryParam("uuid")
             @NotEmpty
             @Pattern(regexp = "[0-9a-fA-F]{8}-([0-9a-fA-F]{4}-){3}[0-9a-fA-F]{12}")
                     String uuid
     ) {
+        if (!isUserPermitted(userRole, new String[]{"admin"})) return Response.status(401).build();
+
         int status = 200;
 
         if (!DataHandler.deleteAirport(uuid)) {
@@ -118,8 +126,10 @@ public class AirportService {
     @Path("create")
     @Produces(MediaType.TEXT_PLAIN)
     public Response addOneAirport(
+            @CookieParam("userRole") String userRole,
             @Valid @BeanParam Airport airport
     ) {
+        if (!isUserPermitted(userRole, new String[]{"admin"})) return Response.status(401).build();
 
         if (airport.getAirportUUID() == null) {
             airport.setAirportUUID(UUID.randomUUID().toString());
@@ -140,11 +150,14 @@ public class AirportService {
     @Path("update")
     @Produces(MediaType.TEXT_PLAIN)
     public Response updateOneAirport(
+            @CookieParam("userRole") String userRole,
             @Valid @BeanParam Airport airport,
             @FormParam("airportUUID")
             @NotEmpty
             @Pattern(regexp = "[0-9a-fA-F]{8}-([0-9a-fA-F]{4}-){3}[0-9a-fA-F]{12}") String uuid
     ) {
+        if (!isUserPermitted(userRole, new String[]{"admin"})) return Response.status(401).build();
+
         int status = 200;
         Airport oldAirport = DataHandler.readAirportByUUID(uuid);
 
@@ -159,5 +172,15 @@ public class AirportService {
                 .status(status)
                 .entity(status == 404 ? "Not Found" : "Object updated")
                 .build();
+    }
+
+    private boolean isUserPermitted(String userRole, String[] roles) {
+        if (userRole == null || userRole.length() == 0) return false;
+
+        for (int i = 0; i < roles.length; i++) {
+            if (roles[i].equalsIgnoreCase(userRole)) return true;
+        }
+
+        return false;
     }
 }
